Input: synaptics - quick hack for ForcePads
ForcePads are found on HP EliteBook 1040 laptops. They lack any kind of
physical buttons, instead they generate primary button click when uses
presses somewhat hard on the surface of the touchpad. Unfortunately they
also report primary button click whenever there are 2 or more contacts
on the pad, messing up all multi-finger gestures (2-finger scrolling,
multi-finger tapping, etc). To cope with this behavior we introduce a
delay in reporting primary press in case more contacts appear.

This is a hack as we do not detect [yet] whether the device is a
ForcePad or a regular ClickPad.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>